es-6.2.1 kibana-6.2.1
- Version management
Because the versions' number of Elasticsearch and Kibana have become accordant with each other since version 5, the version number indicators in this Dockerfile have now been united as ek_version and defined by ARG statement.

- Introducing ES_TMPDIR (TENTATIVE)
From Elasticsearch 6.2.0, it comes with a script bin/elasticsearch-env that creates a temporary directory and defines ES_TMPDIR if they do not exist. In the script, however, it uses mktemp command to make the directory, which is not defined in POSIX shell, resulting in some issues in the different environment. Alpine Linux cannot deal with the way the script uses mktemp command, so this Dockerfile creates a work directory with a fixed path in advance and defines an environmental variable ES_TMPDIR by ENV statement in the Dockerfile. This is a kind of workarounds and should be rolled back when the part of the script becomes portable.