- Base image change (openjdk:jre-alpine -> node:10.15.2-alpine) From version 7 ELK family (and also some upper versions of 6,) Kibana script adds an argument '--max-http-header-size' to nodejs command. Howevr, Alpine's apk does not provide a newer version of nodejs which supports the argument. This is the main reason I changed the base image to the one of node that comes with the appropriate version of nodejs. Furthermore, Kibana script requires nodejs version 10.15.2 and strictly checks it. So the base image must be node:10.15.2-alpine, not the latest or LTS version.
- New ENV variable (JAVA_HOME)
JAVA_HOME is now set to suppress the warning of Elasticsearch.

- Install nss (TENTATIVE)
The apk package 'openjdk8-jre-base' has a bug of dependency (https://bugs.alpinelinux.org/issues/10126), resulting in failure to start Elasticsearch because libnss3.so is missing. Now nss package is added to address the issue. This is a tentative solusion and will be removed when the bug is fixed.