Upgraded ES/KIBANA to 7.5.1 and applied some remedy for tentative solutions
- Upgrade
ES and Kibana are upgraded to version 7.5.1.

- No nss
The nss package that is installed for a tentative solution is no longer needed. See the commit message in ca077fb1795a74be96b000a474eac40542ed477b for more detail about the nss issue.

- No ES_TMPDIR
The environmental variable ES_TMPDIR and mkdir-ing for it are no longer needed since the original ES script stops to use mktemp command in it. See the commit message in 0a7871d053d7e7b00ac8bb42ef8fdcc36b605e96 for more detail about the mktemp issue.

- No JDK bundled
Elasticsearch (OSS, no JDK) is now used to save the image size.

- (TODO) jre8
Now using jre8, but ES is warning it requires jre11 or 13 in the future releases. The alpine running in the base image does not provide those packages.